Remove one animation, make other durations user-definable
jQuery provides some default animation duration constants that
can be changed or turned off by the user if he want's. Constant
ms times can not.

This removes only one animation when the drop-down menu to select
(for example) the calendar model is opened and closed. It does not
make much sense to animate this since no other similar drop-down
is animated.