Re-created $.ui.suggester
Along with removing dependencies on jQuery.ui.autocomplete and jQuery.ui.menu, the most
complicated code structures have been removed. In addition, most of the suggester's
options and dependencies have been removed in order to simplify the widget even more.
With the suggester becoming more lightweight, a preferred way of using is is to create
a widget deriving from the suggester (see jQuery.ui.suggestCommons).