Don't reset the input value on suggester blur
this._term can be an old value which might significantly
differ differ from the current form value.

This causes:
If one adds a sitelink, quickly pastes a page title
and hits enter this code block will set the value of the
input back to the one before the paste (empty string).