Add fallbacks for critical messages
This is not a fix for bug 68033. But it's good to have fallbacks for
the most critical messages in case something is wrong. The "more"
message is the message next to the expand arrow in all experts.
The calendar model can very easily show the untranslated calendar
model name. In both cases the alternative was to show "null" which
is not helpful, doesn't show up in any log but makes the UI unusable
for no good reason.

Bug: 68033