Remove rotation animation from listrotator
This is a first step, so each patch is as small as possible and
always leaves the UI in a non-broken and non-broken looking state.

This removes the rotation animation when you click on an other geo
precision or globe, or a date precision or calendar model. The number,
position and behavior of all UI elements is untouched. The only
difference is that clicks are now instant.

Bug: T109455
Bug: T109457
Change-Id: I19e628ae9072935bcf6299a8053f2544a9817117