Slightly faster inputextender's expand animation
Note that there are two phases of animation: First, the expert fades
in (opacity is animated from 0 to 1). This happens very fast, 150ms.
When this is done, the stuff below the expert moves down. This happens
with the default "fast" animation speed (200ms by default).

I was under the impression that fixed numbers can not be turned off,
but this was wrong. $.fx.off=true turns them all off, at least in this
case where it's a jQuery controlled animation.

Bug: T103834
Change-Id: I8657e178d2a0b355eabc504f95a1ab883644d3c4