Refactor entity search caching
The cache is not meant to cache results for different search terms,
but to cache previous pages when you click the "more" button.

This code is mostly a copy from entityselector. This patch is a copy
from If12e00e and I68d30da.

Note that the caching in this file is currently unused because there
is no more button. I want to fix this later, but not remove the
caching because of this.

Change-Id: I6826d9b75fd70e358cadf97ff7b7143f4d31eb47