Fix broken tests and release 0.20.1
Jenkins is not running the QUnit tests here! You have this do locally
to verify this change (or trust me that I did it already).

This also changes all remaining numbers of assertions into
assert.expect( … ). I don't know why this was not changed in
I15fc0b7. Was this a mistake?

Bug: T170515
Change-Id: Idbbf5376ef98bd0f3c30d1f174e26a500b81b306