Revert "Fix malformed input error"
This reverts commit def53ede5fabce976c7b952257c22e9fe1f77d4f.

Reason for revert: we decided to fix T261071 by not using the JS
StringParser (see I6548c0149a); consequently, sanitizing the value in
JS in neither necessary nor a good idea. We should send the original
string to the PHP StringParser and let that take care of trimming (and
also Unicode normalization, which this sanitization didn’t attempt).

Bug: T261071
Change-Id: I50154171ff27c72b292673a81e5e21bfbdf36c1c