Translate time precision
This adds 15 interface messages for localized time precision levels,
which previously had been hard-coded to English. (The year1k message is
changed from “millenium”, which Wiktionary [1] describes as misspelling
or archaic spelling, to the standard “millennium”.)

The messages are named *-timeinput-*, which is inconsistent with the
calendar messages (*-timevalue-*), but consistent with the globe
coordinate precision messages (*-globecoordinateinput-*); I also think
“input” makes more sense than “value”, since as far as I understand we
only use these messages for the input, not to format the value
(formatting is done through API calls to the PHP code, though possibly
that wasn’t the case when the *-timevalue-calendar-* messages were
introduced).

[1]: https://en.wiktionary.org/wiki/millenium#English

Bug: T269231
Change-Id: Id9d9183361f0d436ac2e24f2e3c169e6d926bbeb