Move autoexpand styles into CSS file
This way, the `left` rule can be flipped by CSSJanus into `right`,
avoiding the issue where the input would be placed in the visible part
of the page and produce a scroll bar.

Note that this requires a corresponding change to Wikibase in order to
add the new file to the jquery.inputautoexpand ResourceLoader module.

(There are probably better ways to implement the “measuring” behavior in
general, but this should be okay for now.)

Bug: T321441
Change-Id: I7bf03ceeda4911e7d7d398043d9da073b8ae7d7e