Use codex token for background color for valueview input
Use the codex token `background-color-base` for the background of
the valueview input instead of a hard-coded colour value. We use
CSS variables here rather than porting this file to LESS since all
other stylesheets in this repo are CSS (we can think about porting
at a later date).

At the same time, update the border highlight for the focussed
input to match our design. This overrides the browser's default
highlight colour for active inputs, which is considered an
accessibility feature.

`.valueview-disabled` still has some hard-coded colour values, but
we do not appear to use that functionality in Wikibase.

Bug: T389349
Change-Id: Iaaa8749d6b328925ec68628dfa9a239cfe89e54b